+++ crates/virtual-dom-rs/Cargo.toml
@@ -9,11 +9,8 @@ repository = "https://github.com/chinedufn/percy"
 documentation = "https://chinedufn.github.io/percy/api/virtual_dom_rs/"
 edition = "2018"
 
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-js-sys = "0.3"
-
 [dependencies]
-
+js-sys = "0.3"
 wasm-bindgen = {version = "0.2", features = ["default", "nightly"]}
 
 [dependencies.web-sys]
